Cherry-Picking Refactor to package
